requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.16
django-datatables-view==1.20.0
django-localflavor==4.0
django-stdimage==5.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==28.4.1
inflection==0.5.1
packaging==24.1
Pillow==9.1.0
python-dateutil==2.9.0.post0
python-stdnum==1.19
pytz==2023.3.post1
PyYAML==6.0.2
six==1.16.0
sqlparse==0.4.4
style==1.1.0
typing_extensions==4.9.0
tzdata==2023.4
update==0.0.1
uritemplate==4.1.1